Grad 657 update upgrade plan

+++ course-service/src/main/java/com/example/courseservice/dto/response/course/DetailCourseResponse.java
@@ -27,8 +27,8 @@ public class DetailCourseResponse {
     String unitPrice;
     UUID userUid;
     int lessonCount;
-    float averageRating;
-    int reviewCount;
+    Double averageRating;
+    Integer reviewCount;
     boolean isUserEnrolled;
     UUID latestLessonId;
     Float progressPercent;
